+++ tests/test_get_collisions.py
@@ -2,13 +2,13 @@
 
 import pytest
 import subprocess as sub
-from bin import get_collisions
+from bin import list_collisions
 
 def test_noargs():
-    sub.call(get_collisions.__file__, shell=True)
+    sub.call(list_collisions.__file__, shell=True)
     
 def test_oor():
-    sub.call([get_collisions.__file__, "-o"], shell=True)
+    sub.call([list_collisions.__file__, "-o"], shell=True)
 
 if __name__ == "__main__":
     pytest.main()